style: :truck: rename composables to unified convention
