+++ src/components/ClientTable.vue
@@ -13,8 +13,8 @@ import ClientModal from '@/components/ClientModal.vue'
 import type { TableColumn } from '@nuxt/ui'
 import { h, resolveComponent } from 'vue'
 import type { Row, SortingState, VisibilityState } from '@tanstack/table-core'
-import { showModal } from '@/composables/showModal'
-import { deleteClient, updateClient } from '@/composables/apiCalls'
+import { showModal } from '@/composables/modalUtils'
+import { deleteClient, updateClient } from '@/composables/apiUtils'
 import { useClientStore } from '@/store/client'
 import { storeToRefs } from 'pinia'
 import ConfirmationModal from '@/components/ConfirmationModal.vue'
